Make unittests optional and disable by default, rename cmake options
The unittests contain hardcoded paths, so they are pretty much expected
to fail on most machines. So these are not optional and have to be
enabled explicitly.

The cmake options to enable the build of the generator application as
well as the unittests not have a LIBLDPC prefix to allow using them
as a subproject.